FIX: Rename the query method by search.
For me it was confusing to have a class Query and have a client with a method with the same name,that is doing actually the request to search document matching the given query build with the class Query, that's the reason why I made this change. Also the word search is more explicit since that's the main feature of solr.