Fix: #71 Introducing json-bigint to properly handle the too-big-number issue in classic json parsing.
Adding the test-cases that show the need for this in regular solr fields of type long (_l suffix)
and the fact that the solr 4.x optimistic locking feature (ie the  _version_ field) needs this badly.

Caveat: json-bigint is currently more restrictive on json parsing then the classic JSON.parse
Namely: it renders an error on duplicate key fields in the json. Since the response-header of the ping
in solr introduces such a duplicate key this commit chooses to force omitHeader=true for the ping request.

This can be reconsidered if json-bigint would allow for a more lenient parse() in line with the classic one.